requirements.txt
Pillow==3.2.0
wheel==0.24.0
